Fix user add tags not update after deleting tag
